requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Speech Recognition
openai-whisper>=20230314
speechrecognition>=3.10.0
pyaudio>=0.2.13

# Machine Translation
sacrebleu>=2.3.1
sentencepiece>=0.1.99
tokenizers>=0.13.3

# Text-to-Speech
gTTS>=2.3.2
pyttsx3>=2.90
TTS>=0.15.6

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.1
pydub>=0.25.1

# Web Interface
gradio>=3.35.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
requests>=2.31.0
python-dotenv>=1.0.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Cloud/MLOps (optional)
boto3>=1.28.0
google-cloud-storage>=2.10.0
wandb>=0.15.0
mlflow>=2.5.0
